Renaming of Pages and Components
