+++ frontend/src/Components/DishListItem/DishListItem.test.js
@@ -1,9 +1,9 @@
 import { render } from '@testing-library/react'
-import ListItem from './ListItem'
+import DishListItem from './DishListItem'
 
-describe('ListItem', () => {
+describe('DishListItem', () => {
   it('renders correctly', () => {
-    const { container } = render(<ListItem title="My Dish" />)
+    const { container } = render(<DishListItem title="My Dish" />)
     expect(container.firstChild).toMatchSnapshot()
   })
 })
