+++ frontend/src/Pages/DishesAllPage.js
